split into two py scripts
